remove dupes from classhashmap
