+++ src/deckyPatches/UnminifyMode.tsx
@@ -1,34 +1,5 @@
-import { classMap } from "decky-frontend-lib";
+import { classHashMap, initializeClassHashMap } from "./ClassHashMap";
 import { getRootElements } from "./SteamTabElementsFinder";
-
-const classHashMap = new Map<string, string>();
-
-function initializeClassHashMap() {
-  const withoutLocalizationClasses = classMap.filter((module) => Object.keys(module).length < 1000);
-
-  const allClasses = withoutLocalizationClasses
-    .map((module) => {
-      let filteredModule = {};
-      Object.entries(module).forEach(([propertyName, value]) => {
-        // Filter out things that start with a number (eg: Breakpoints like 800px)
-        // I have confirmed the new classes don't start with numbers
-        if (isNaN(Number(value.charAt(0)))) {
-          filteredModule[propertyName] = value;
-        }
-      });
-      return filteredModule;
-    })
-    .filter((module) => {
-      // Some modules will be empty after the filtering, remove those
-      return Object.keys(module).length > 0;
-    });
-
-  allClasses.forEach((module: Record<string, string>) => {
-    Object.entries(module).forEach(([propertyName, value]) => {
-      classHashMap.set(value, propertyName);
-    });
-  });
-}
 
 export function unminifyElement(element: Element) {
   if (element.classList.length === 0) return;
